pyproject.toml
[project]
name = "image-object-coordinate-finder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
]
